+++ cpp/tests/utilities/tuple_vectors.h
@@ -20,6 +20,7 @@
 #include <type_traits>
 #include <iostream>
 #include <cstdlib>
+#include <iterator>
 
 template <typename... T>
 using VTuple = std::tuple<std::vector<T>...>;